Reorganise components/room
