+++ Web/src/components/create-room/index.tsx
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import BigButton from 'components/common/BigButton';
 import Loading from 'components/common/Loading';
-import CommContext from 'components/room/CommContext';
-import useCommHooks from 'components/room/hooks/useCommHooks';
+import CommContext from 'components/room/comm/CommContext';
+import useCommHooks from 'components/room/comm/useCommHooks';
 import NumPlayers from './NumPlayers';
 import SocialShare from './SocialShare';
 import QrCode from './QrCode';
