--- Web/src/components/room/PresenceState.ts
@@ -1,5 +1,0 @@
-import Meta from './Meta';
-
-export default interface PresenceState {
-  [clientId: string]: { metas: Meta[] }
-}
