+++ Web/src/components/room/comm/Errors.ts
@@ -1,0 +1,11 @@
+export enum CommError {
+  NoMorePin = 'NoMorePin',
+  Timeout = 'Timeout',
+  Other = 'Other',
+}
+
+export enum PushError {
+  NoChannel = 'NoChannel',
+  Timeout = 'Timeout',
+  Other = 'Other',
+}
