--- Web/src/components/room/hooks/ChannelResponse.ts
@@ -1,7 +1,0 @@
-enum ChannelResponse {
-  OK = 'ok',
-  ERROR = 'error',
-  TIMEOUT = 'timeout'
-}
-
-export default ChannelResponse;
