--- Web/src/components/room/hooks/ErrorCause.ts
@@ -1,6 +1,0 @@
-enum ErrorCause {
-  Timeout = 'timeout',
-  Other = 'other'
-}
-
-export default ErrorCause;
